renombrado a enrollment-card, clases css compartidas con la +course-mini-card para la +sección explore con el fin de reutilizado +fix en el doble disparo de saveLesson, +renombrado de metodo onSumbit en formLesson

+++ src/app/lesson/components/form-lesson/form-lesson.component.ts
@@ -33,7 +33,7 @@ export class FormLessonComponent {
 
   constructor() { }
 
-  public onSubmit = () => {
+  public onSaveLesson = () => {
     
     this.lessonFormState.lessonForm.markAllAsTouched();
 
@@ -52,34 +52,38 @@ export class FormLessonComponent {
       lessonDto.lesson_number = this.lessonFormState.lessons().at(-1)?.lesson_number ?? 0;
 
       if( this.lessonFormState.mediaFile() != null ){
-        this.fileService.updateFile( this.folder, this.lessonFormState.mediaFile()! )
+        return this.fileService.updateFile( this.folder, this.lessonFormState.mediaFile()! )
                           .subscribe( response => {
-                              console.log(response);
                               lessonDto.id_file = response.id;
                               console.log(lessonDto);
                               return this.lessonService.saveLesson( lessonDto ).subscribe();
                           });
       }
-      return this.lessonService.saveLesson( lessonDto ).subscribe();
+      else{
+        return this.lessonService.saveLesson( lessonDto ).subscribe();
+      }
     }
     return;
   }
 
   public onDeleteLesson = ( id : string ) => {
-    if( this.courseFormState.course()?.id_owner === this.authService.id() ){
-      if( this.lessonFormState.lessonSelected()?.file.id_file ){
-        this.fileService.deleteCourseThumbnail( this.courseFormState.course()?.id! ).subscribe()
-      }
-      this.lessonService.deleteLesson( id )
-                            .subscribe( ( isLessonDeleted ) => {
-                                if( isLessonDeleted ) {
-                                  this.router.navigateByUrl('/');
-                                  return;
-                                }     
-                            } );
+    if( id ){
+        if( this.courseFormState.course()?.id_owner === this.authService.id() ){
+          if( this.lessonFormState.lessonSelected()?.file.id_file ){
+            this.fileService.deleteCourseThumbnail( this.courseFormState.course()?.id! ).subscribe()
+          }
+          this.lessonService.deleteLesson( id )
+                                .subscribe( ( isLessonDeleted ) => {
+                                    if( isLessonDeleted ) {
+                                      this.router.navigateByUrl('/');
+                                      return;
+                                    }     
+                                } );
+        }
     }
-  }
-    
+    else{
+    }
+  }    
 
 
 }
